refactor: Move is_proxy_skipped to utils
